GPU (FreeBSD): moves DRM detection to global space for OpenBSD use
